+++ packages/shared/src/components/post/smartPrompts/SmartPrompt.tsx
